requirements.txt
py-cord
tinydb
pyyaml
python-dotenv
python-dateutil
pytz

# py-cord speedups
aiodns
brotlipy
# cchardet
orjson